// src/types/menu-tab.enum.ts
export type menuList = mainTabs.MAIN | mainTabs.OPPORTUNITIES | mainTabs.WISHLIST | mainTabs.WISHLIST_UNIVERSITY | mainTabs.WISHLIST_OPPORTUNITIES;
export enum mainTabs {
  MAIN = 'main',
  OPPORTUNITIES = 'opportunities',
  WISHLIST = 'wishlist',
  WISHLIST_UNIVERSITY = 'wishlist_university',
  WISHLIST_OPPORTUNITIES = 'wishlist_opportunities',
  TRACKER = 'tracker',
  PROFILE = 'profile',
}